requirements.txt
Django >=3.0.2,<4.0
jinja2
pytest-django
pytest-factoryboy
